% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{boot_eg}
\alias{boot_eg}
\title{?}
\usage{
boot_eg(data, registry_years, registry_start_year, age, sex, N_boot = 1000)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{age}{Age of example patient.}

\item{sex}{Sex of example patient ("Male" or "Female").}

\item{N_boot}{Number of replicates.}
}
\value{
?
}
\description{
?
}

