% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{counted_prevalence}
\alias{counted_prevalence}
\title{Count prevalence from registry data.}
\usage{
counted_prevalence(entry, eventdate, status, start = NULL,
  num_reg_years = NULL, indexdate = NULL)
}
\arguments{
\item{entry}{Vector of diagnosis dates for each patient in the registry.}

\item{status}{Vector of binary values indicating if an event has occurred for each patient in the registry.}

\item{start}{Date from which incident cases are included.}

\item{num_reg_years}{Integer representing the number of complete years of the registry for which incidence is to be calculated.}

\item{indexdate}{Index date at which to estimate prevalence.}

\item{events}{Vector of dates corresponding to the indicator variable.}
}
\value{
A count of prevalence at the index date subdivided by year of diagnosis and inclusion in the registry.
}
\description{
Count prevalence from registry data.
}
\examples{
counted_prevalence(registry_data$entrydate,
                   registry_data$eventdate,
                   registry_data$status,
                   indexdate = "2013-01-30",
                   start="2004-01-30", num_reg_years=8)
}

