% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_plots.R
\name{cumulative_incidence}
\alias{cumulative_incidence}
\title{Generate smoothed cumulative incidence function, inspect deviations in the registry data
and compare with the cumulative incidence for constant diagnosis rate.}
\usage{
cumulative_incidence(entry, start = NULL, num_reg_years = NULL, df = 6)
}
\arguments{
\item{entry}{Vector of diagnosis dates for each patient in the registry in the format YYYY-MM-DD.}

\item{start}{Date from which incident cases are included in the format YYYY-MM-DD,
defaults to the earliest entry date.}

\item{num_reg_years}{The number of years of the registry for which incidence is to be calculated,
defaults to using all available complete years.}

\item{df}{Integer representing degrees of freedom for the smoothening function.}
}
\value{
Object containing incidence for each year of the registry, the cumulative incidence functions and corresponding deviations.
}
\description{
The plot generated by the example shows the cumulative number of diagnoses in the registry
by day in black. This is compared to the red line, which indicates what cumulative diagnosis
would look like if the rate of diagnosis was constant. The green line is a smooth fitted to
the actual cumulative diagnosis data.
}
\examples{
c_inc <- cumulative_incidence(registry_data$entrydate, start = "2004-01-30", num_reg_years = 9)
ordered_diagnoses <- c_inc$ordered_diagnoses
cum_inc <- c_inc$cumulative_incidence
smooth <- c_inc$smooth

plot(ordered_diagnoses, cum_inc, pch=20,
cex=0.7, xlab="days", ylab="cumulative diagnoses")

abline(a=0, b=length(ordered_diagnoses)/ordered_diagnoses[length(ordered_diagnoses)],
col="red", lwd=2)

lines(smooth, col="green", lwd=2)

# The following plot shows the deviation of the raw data from the fitted smooth by day:
plot(ordered_diagnoses, seq(length(ordered_diagnoses)) - predict(smooth, ordered_diagnoses)$y,
type="l", xlab="days", ylab="deviation from smooth")
}

