% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curemodel.R
\name{cure_model_wrapper}
\alias{cure_model_wrapper}
\title{Builds a cure model with an associated population mortality table}
\usage{
cure_model_wrapper(formula, data, dist = "weibull", pop_data = NULL,
  link = "logistic", mixture = T, ...)
}
\description{
Builds a cure model with an associated population mortality table
}
