% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{daily_survival_rate}
\alias{daily_survival_rate}
\title{Population survival rate.}
\usage{
daily_survival_rate(data, sex)
}
\arguments{
\item{data}{A dataset of population survival data by year.}

\item{sex}{Can be "Males" or "Females".}
}
\value{
An estimate of the survival rate by day for ages <=100 years.
}
\description{
Population survival rate.
}
\examples{
daily_survival_males <- daily_survival_rate(population_data_mx, sex = "Males")
daily_survival_females <- daily_survival_rate(population_data_mx, sex = "Females")
}

