% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{dfr_sc}
\alias{dfr_sc}
\title{?}
\usage{
dfr_sc(age, sex, wb)
}
\arguments{
\item{age}{Age of example patient.}

\item{sex}{Sex of example patient ("Male" or "Female").}

\item{wb}{Weibull model fitted to the data.}
}
\value{
?
}
\description{
?
}

