% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedcure.R
\name{fixed_cure}
\alias{fixed_cure}
\title{Fixed cure model where survival reverts to population mortality
at a specified timepoint.}
\usage{
fixed_cure(formula, data, dist, time, pop_data)
}
\arguments{
\item{formula}{Survival formula}

\item{data}{Dataset}

\item{dist}{Distribution to be used}

\item{time}{Time-limit at which a patient is considered cured. Uses
same time-scale as survival time in \code{data}.}

\item{pop_data}{A data frame comprising population survival as a daily rate for 36525 days (100 years).
It can also be stratified by other variables that are found in the survival \code{formula} for this model,
such as sex.}
}
\value{
An object of class \code{fixedcure} that can be passed
  into \code{\link{prevalence}}.
}
\description{
Fixed cure model where survival reverts to population mortality
at a specified timepoint.
}
