% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvcure.R
\name{flexsurvcure_population}
\alias{flexsurvcure_population}
\title{Builds survival models for diseases with cured fractions using population mortality tables}
\usage{
flexsurvcure_population(formula, data, daily_survival = NULL,
  population_covariates = NULL, dist = "weibull", link = "logistic",
  mixture = TRUE, ...)
}
\arguments{
\item{formula}{Formula specifying survival model as used by \code{\link{flexsurvcure}{flexsurvcure}}.}

\item{data}{A data frame in which to find variables supplied in
\code{formula}.  If not given, the variables should be in the working
environment.}

\item{daily_survival}{A data frame comprising population survival as a daily probability for as long as possible,
ideally 100 years (36525 days).
Defaults to using UK mortality from the \code{UKmortalitydays} data set.
It \strong{must contain columns 'age' and 'surv'}, providing the age (in days) and survival
probability at that age respectively.
It can also be stratified by other variables that are found in the survival \code{formula} for this model,
such as sex.}

\item{population_covariates}{A character vector containing fields to stratify population survival by in addition to
age, as descripted in \code{Details} below. These \strong{must} be the names of columns in both \code{data} and
\code{daily_survival}}

\item{dist}{A string representing one of the built-in distributions of flexsurv.
\code{Surv(time, dead) ~ age + treat, anc = list(shape = ~ sex + treat)}}

\item{link}{A string representing the link function to use for estimation of the
cure fraction.  Defaults to logistic.}

\item{mixture}{optional TRUE/FALSE to specify whether a mixture model should be fitted.  Defaults to TRUE.}

\item{...}{other arguments to be passed to \code{\link{flexsurvreg}}.}
}
\value{
An object of class \code{flexsurvcure}.
}
\description{
Fits mixture and non-mixture cure models using \code{flexsurvcure}. Please read the detailed description below
  for how to use this model.
}
\details{
To model population survival, population mortality tables are required, as specified by the \code{daily_survival}
argument. If not provided, then the default population mortality is that of the UK population, which goes up
to 100 years of age. If a simulated individual has expected lifespan longer than the maximum age in the mortality table
then they are estimated to have died at this age limit,
which is why it is advantageous to provide as many accurate survival probabilities as possible.

Due to the linking with the registry data and the ability for user-specified mortality tables, there are stricter
requirements on the survival models used in cure models than elsewhere. For example, the time-scale of the
survival model specified in \code{formula} \strong{must be in days} so that it matches up with the mortality tables.
Likewise, \strong{age in years must be included as a covariate} in the survival model
}
