% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{functional_form_age}
\alias{functional_form_age}
\title{Inspect whether a non-linear transform of age is appropriate.}
\usage{
functional_form_age(form, data, df = 4, plot_fit = T)
}
\arguments{
\item{form}{Formula where the LHS is represented by a standard \code{Surv} object, and the RHS
provides the name of the variable containing age in the supplied dataset.}

\item{data}{A data frame with the corresponding column names provided in \code{form}.}

\item{df}{The desired degrees of freedom for the cubic spline..}

\item{plot_fit}{Whether to plot the relationship between age and relative hazard.}
}
\value{
An \code{rms::cph} object containing the Cox model fit between a cubic spline transform
of age and the survival outcomes.
}
\description{
Inspect whether a non-linear transform of age is appropriate.
}
\examples{
data(prevsim)

functional_form_age(Surv(time, status) ~ age, prevsim)

functional_form_age(Surv(time, status) ~ age, prevsim, df=2)

functional_form_age(Surv(time, status) ~ age, prevsim, df=2, plot_fit=F)

}

