% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Calculate absolute incidence from registry data.}
\usage{
incidence(data, registry_years, registry_start_year, registry_end_year)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of registry.}

\item{registry_start_year}{Ordinal defining the first year of registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of registry data to be used.}
}
\value{
Vector of absolute incidence values for each of the specified years of a registry.
}
\description{
Calculate absolute incidence from registry data.
}
\examples{
registry_years <- c("2004-09-01", "2005-09-01", "2006-09-01", "2007-09-01", "2008-09-01")
incidence(load_data(registry_data), registry_years, registry_start_year = 1, registry_end_year = 4)
}

