% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{load_data}
\alias{load_data}
\title{Load registry data.}
\usage{
load_data(data, age_initial = "age", date_initial = "DateOfDiag",
  indicator = "status", survival_time = "stime", date_event = "EventDate",
  indicator_censored_at_index = "status2", sex = "Both",
  age_division = "None", age_cut = 20)
}
\arguments{
\item{data}{A registry dataset of patient cases. Cases with missing values will be removed.}

\item{age_initial}{The column name representing age at data_initial.}

\item{date_initial}{The column name representing the date cases are added to the registry.}

\item{indicator}{The column name representing a binary variable where 1 is event has occurred and 0 is event has not occurred, or the case is censored.}

\item{survival_time}{The column representing time to event or censorship in days.}

\item{date_event}{The column name representing the date an event or censorship has occurred.}

\item{indicator_censored_at_index}{The column name representing a binary variable where 1 is event has occurred and 0 is event has not occurred, or the case is censored, by an index date of interest.}

\item{sex}{Can be "Both" for all cases, "Males" or "Females". Please note that the column name must be exactly "sex" to enable correct filtering.}

\item{age_division}{Can be "None", "Adult" or "Paediatric".}

\item{age_cut}{A number to define age cut-off in adult or paediatric analyses.}
}
\value{
A dataset restricted to a population of desired age and sex. Selected age, sex, date and indicator columns are named appropriately for use in subsequent functions within this package.
}
\description{
Load registry data.
}
\examples{
load("../data/registry_data.rda")
summary(load_data(registry_data))
}

