% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{mean_incidence_rate}
\alias{mean_incidence_rate}
\title{Calculate average incidence rates per one hundred thousand over the specified number of years, with
confidence intervals.}
\usage{
mean_incidence_rate(entry, population_size, start = NULL,
  num_reg_years = NULL, precision = 2, level = 0.95)
}
\arguments{
\item{entry}{Vector of diagnosis dates for each patient in the registry in the format YYYY-MM-DD.}

\item{population_size}{The size of the population at risk.}

\item{start}{Date from which incident cases are included in the format YYYY-MM-DD, defaults to the earliest entry date.}

\item{num_reg_years}{The number of years of the registry for which incidence is to be calculated, defaults to using all available complete years.}

\item{precision}{The number of decimal places required.}

\item{level}{The desired confidence interval width.}
}
\value{
A list with the following values:

\item{absolute}{Overall incidence for the period of interest as a single double}
\item{per100K}{Incidence for the period of interest per one hundred thousand}
\item{per100K.lower}{Lower bounds of the specified confidence level on the per one hundred thousand estimate}
\item{per100K.upper}{Upper bounds of the specified confidence level on the per one hundred thousand estimate}
}
\description{
Calculate average incidence rates per one hundred thousand over the specified number of years, with
confidence intervals.
}
\examples{
data(prevsim)

mean_incidence_rate(prevsim$entrydate, population_size=3.5e6, start='2004-01-01', num_reg_years=8)
mean_incidence_rate(prevsim$entrydate, population_size=3.5e6)
mean_incidence_rate(prevsim$entrydate, population_size=3.5e6, precision=3)
mean_incidence_rate(prevsim$entrydate, population_size=3.5e6, start='2004-01-01', num_reg_years=8, level=0.99)
}

