% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{meanIR}
\alias{meanIR}
\title{Convert absolute incidence values to per 100,000 population values.}
\usage{
meanIR(data, registry_years, registry_start_year, registry_end_year,
  population_size, precision = 2)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of the registry data to be used.}

\item{population_size}{A number representing the size of the population at risk.}

\item{precision}{The number of decimal places required.}
}
\value{
A vector containing mean incidence/year and incidence rate/100,000/year with confidence
        intervals.
}
\description{
Convert absolute incidence values to per 100,000 population values.
}
\examples{
incidence_rates <- meanIR(load_data(registry_data), registry_years, registry_start_year = 1,
registry_end_year = 4, population_size = 3500000)
}

