% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{n_year_estimates}
\alias{n_year_estimates}
\title{Calculate predicted prevalence for a given number of years.}
\usage{
n_year_estimates(object, num_years_to_estimate, population_size, level = 0.95,
  precision = 2)
}
\arguments{
\item{object}{A \code{prevalence} object.}

\item{num_years_to_estimate}{Integer representing number of years prevalence is to be calculated for.}

\item{population_size}{Integer corresponding to the size of the population.}

\item{level}{Double representing the desired confidence interval width.}

\item{precision}{Integer representing the number of decimal places required.}
}
\value{
A list with the following attributes:
\item{absolute.prevalence}{Absolute prevalence for the specified time frame.}
\item{per100K}{Prevalence rates per one hundred thousand in the specified population.}
\item{per100K.lower}{Lower confidence bounds on the estimate.}
\item{per100K.upper}{Upper confidence bounds on the estimate.}
}
\description{
Calculate predicted prevalence for a given number of years.
}
\examples{
data(prevsim)

prev <- prevalence(Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate),
                   data=prevsim, num_years_to_estimate = 10, start = "2005-09-01",
                   num_reg_years = 8, cure_days = 5*365)

n_year_estimates(prev, num_years_to_estimate = 3, population_size = 1.7e6)

n_year_estimates(prev, num_years_to_estimate = 5, population_size = 1.7e6)

# Maximum number of years to estimate for is 10 as this is the number of years the prev object
# was assigned to estimate.
n_year_estimates(prev, num_years_to_estimate = 10, population_size = 1.7e6, level=0.99)
}

