% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{n_year_estimates_current}
\alias{n_year_estimates_current}
\title{Calculate predicted prevalence for a given number of years.}
\usage{
n_year_estimates_current(num_years_to_estimate, registry_start_year,
  registry_end_year, the_samples, by_year, population_size)
}
\arguments{
\item{num_years_to_estimate}{Number of years prevalence is to be calculated for.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of the registry data to be used.}

\item{the_samples}{Bootstrapped samples used to gauge the sampling variation in the estimates.}

\item{by_year}{Predicted prevalence by year of the registry.}

\item{population_size}{A number corresponding to the size of the population.}
}
\value{
An estimate of prevalence, in total and /100,000 with confidence intervals.
}
\description{
Calculate predicted prevalence for a given number of years.
}
\examples{
n_year_estimates_current(num_years_to_estimate = 3, registry_start_year = year1, registry_end_year = lastyear,
                 the_samples = by_year_male_samples,
                 by_year = by_year_male, population_size = 1700000)
}

