% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{plot.incidence}
\alias{plot.incidence}
\title{Plots a comparison between the smoothed daily incidence function and actual incidence.}
\usage{
\method{plot}{incidence}(object, level = 0.95)
}
\arguments{
\item{object}{An \code{incidence} object.}

\item{level}{The desired confidence interval width.}
}
\value{
Plot of incidence rate, confidence interval and smoothed incidence function as a side-effect.
}
\description{
This function generates a plot from the cumulative incidence object. The incidence rate per
year of the registry is shown in red. Mean incidence rate is shown as a solid blue line,
with the confidence interval shown in dashed blue lines. The smooth fitted to the
cumulative incidence data is shown in green.
}
\examples{
data(prevsim)

inc <- incidence(previm$entrydate, population_size=1e6, start = "2004-01-30", num_reg_years = 9)

plot(inc)

}

