% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{plot_incidence_fit}
\alias{plot_incidence_fit}
\title{Visualise incidence Poisson assumption.}
\usage{
plot_incidence_fit(object, N_sim = 1000, level = 0.95,
  samples_per_bin = 10, max_bins = 200)
}
\arguments{
\item{object}{An \code{incidence} object.}

\item{N_sim}{Number of draws from a homogeneous Poisson process.}

\item{level}{The desired confidence interval width.}

\item{samples_per_bin}{Number of samples per bin.}

\item{max_bins}{Maximum number of bins.}
}
\value{
Plot of the smoothed incidence function and corresponding deviations.
}
\description{
Plots a figure detailing the deviation of incidence rate from an homogeneous
Poisson process using simulation.
}
\details{
This function generates a plot where the red line is the deviation of
cumulative diagnoses from the fitted smooth taken from the incidence object.
In order to ascertain if the deviation is within reasonable bounds,
simulation is used. The grey lines represent N draws from a uniform
distribution between 0 and the last day of diagnosis, where N is the number
of incident cases, plotted as deviations from a smooth fitted to them. The
blue lines are 95\% confidence intervals drawn from the simulated data.
}
\examples{
data(prevsim)

inc <- incidence(prevsim$entrydate, 1e6, start = "2004-01-30",
                 num_reg_years = 9)

plot_incidence_fit(inc)
}

