% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_km}
\alias{plot_km}
\title{?}
\usage{
plot_km(data, registry_years, registry_start_year, age, sex, limits,
  colour = "green")
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{age}{Age of example patient.}

\item{sex}{Sex of example patient ("Male" or "Female").}

\item{limits}{?}

\item{colour}{?}
}
\value{
?
}
\description{
?
}

