% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_pop_Su}
\alias{plot_pop_Su}
\title{?}
\usage{
plot_pop_Su(data, age, sex, the_length = 3000)
}
\arguments{
\item{data}{Population survival dataset.}

\item{age}{Age of example patient.}

\item{sex}{Sex of example patient ("Male" or "Female").}

\item{the_length}{Number of days to predict.}
}
\value{
?
}
\description{
?
}

