% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_Su}
\alias{plot_Su}
\title{?}
\usage{
plot_Su(age, sex, the_length = 3000, wb)
}
\arguments{
\item{age}{Age of example patient.}

\item{sex}{Sex of example patient ("Male" or "Female").}

\item{the_length}{Number of days to predict.}

\item{wb}{Weibull model fitted to the data.}
}
\value{
?
}
\description{
?
}

