% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_incidence.R
\name{poisson_incidence_sim_gg}
\alias{poisson_incidence_sim_gg}
\title{\code{poisson_incidence_sim} adapted for ggplot2.}
\usage{
poisson_incidence_sim_gg(object, N_sim = 1000, level = 0.95,
  samples_per_bin = 10, max_bins = 200)
}
\arguments{
\item{object}{A \code{cincidence} object.}

\item{N_sim}{Number of draws from a homogeneous Poisson process.}

\item{level}{The desired confidence interval width.}

\item{samples_per_bin}{Number of samples per bin.}

\item{max_bins}{Maximum number of bins.}
}
\description{
See \code{\link{poisson_incidence_sim}}.
}

