% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{data}
\name{population_data_mx}
\alias{population_data_mx}
\title{General population survival data.}
\format{A data frame with 1000 rows and 4 variables:
\describe{
 \item{calendar_year}{year}
 \item{sex}{sex, "Males" or "Females"}
 \item{age}{age, in years}
 \item{rate}{death rate, by year}
}}
\usage{
population_data_mx
}
\description{
A dataset containing central rates of mortality of males and females of different ages from
the UK population, obtained from:
\url{http://www.ons.gov.uk/ons/taxonomy/index.html?nscl=Life+Tables#tab-data-tables}.
Adapted from public sector information licensed under the Open Government Licence v3.0.
Data were relabelled according to the mean year of the three-year birth window.
The user is advised to select "population_data" or "population_data_mx" according to their needs.
}
\keyword{datasets}

