% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{population_survival_rate}
\alias{population_survival_rate}
\title{Population survival rate.}
\usage{
population_survival_rate(form, data, max_age = 100)
}
\arguments{
\item{form}{Formula to define...}

\item{data}{Population data frame containing calendar_year, sex, age and rate.}

\item{max_age}{Integer to specify maximum age.}
}
\value{
An estimate of the survival rate by day.
}
\description{
Population survival rate.
}
\examples{
pop_data_mx <- load(population_data_mx)
daily_survival_males <- population_survival_rate(rate ~ age, subset(pop_data_mx, sex==0))
}

