% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_model.R
\name{predict_event_time}
\alias{predict_event_time}
\title{Predicts event times for given individuals.}
\usage{
predict_event_time(object, newdata)
}
\arguments{
\item{object}{The survival object itself}

\item{newdata}{Simulated incident individuals, with the same attributes specified in
\code{extract_covars} and found in the supplied registry data to \code{prevalence}.}
}
\value{
A vector with the length \code{nrow(newdata)} providing the estimated
  event times for these individuals.
}
\description{
This generic method is optional for any survival object used in the main
\code{prevalence} function. If it isn't provided, then the
\code{predict_survival_probability} method must be accessible.
If both methods are available, \code{rprev} will default to using
\code{predict_event_time}, as it more computationally efficient.
However, some models are unable to provide an estimated event time
and so must use the more flexible \code{predict_survival_probability}.
}
