% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{prev_chisq}
\alias{prev_chisq}
\title{Chi squared test between prevalence prediction and observed values in the registry.}
\usage{
prev_chisq(object)
}
\arguments{
\item{object}{A \code{prevalence} object.}
}
\value{
P-value from a chi-squared test of difference between prevalence prediction and counted prevalence at the index date.
}
\description{
Chi squared test between prevalence prediction and observed values in the registry.
}
\examples{
prev_chisq(prevalence_total)
}

