% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{prev_chisq_current}
\alias{prev_chisq_current}
\title{Chi squared test between prevalence prediction and observed values in the registry.}
\usage{
prev_chisq_current(data, registry_years, registry_start_year, registry_end_year,
  by_year)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of the registry data to be used.}
}
\value{
Plots of the smoothed incidence function and corresponding deviations.
}
\description{
Chi squared test between prevalence prediction and observed values in the registry.
}
\examples{
prev_chisq_current(load_data(registry_data), registry_years = registry_years, registry_start_year = registry_start_year, registry_end_year = registry_end_year, by_year = by_year_total)
}

