% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence_by_age}
\alias{prevalence_by_age}
\title{Output extrapolated number of prevalent cases per 10-year age group.}
\usage{
prevalence_by_age(dist, registry_end_year, N_years)
}
\arguments{
\item{dist}{The age distribution of prevalent cases following simulation.}

\item{registry_end_year}{The last complete year of registry data.}

\item{N_years}{The year to make predictions until.}
}
\value{
Predicted number of prevalent cases per 10-year age group.
}
\description{
Output extrapolated number of prevalent cases per 10-year age group.
}
\examples{
result <- prevalence_by_age(dist = post_age_dist_male, registry_end_year = 4,
                 N_years = N_years)
}

