% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\docType{data}
\name{prevsim}
\alias{prevsim}
\title{Simulated patient dataset.}
\format{A data frame with 1000 rows and 6 columns:
\describe{
 \item{time}{(double) time between date of diagnosis and death or censorship, in days}
 \item{status}{(integer) event marker, 1 if patient is deceased and 0 if alive or censored}
 \item{age}{(double) age in years at point of entry into the registry}
 \item{sex}{(integer), 0 for males or 1 for females}
 \item{entrydate}{(character) date of entry into the registry in YYYY-MM-DD format}
 \item{eventdate}{(character) date of death or censorship in YYYY-MM-DD format}
}}
\usage{
prevsim
}
\description{
A dataset in the format of a disease registry, where the outcome being modeled is death
due to the simulated disease. The registry began in March 2013, with 1000 incident cases being
recorded over a period of nearly exactly ten years. The patients are followed up for a further two years
until 17.03.2015, at which point any subjects alive are marked as right censored.
}
\details{
Demographic and disease-specific data required for prevalence estimations are included,
such as sex, age, and dates of entry and event. \code{eventdate} marks the date of the
last known follow-up with the patient, which is either death (\code{status = 1}) or
right-censored (\code{status = 0}).
}
\keyword{datasets}

