% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.incidence}
\alias{print.incidence}
\title{Print the current value of the incidence object.}
\usage{
\method{print}{incidence}(object, ...)
}
\arguments{
\item{object}{...}
}
\value{
...
}
\description{
Print the current value of the incidence object.
}
\examples{
...
}

