% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.prevalence}
\alias{print.prevalence}
\title{Print the current value of the prevalence object.}
\usage{
\method{print}{prevalence}(object, ...)
}
\arguments{
\item{object}{A \code{prevalence} object.}
}
\description{
Print the current value of the prevalence object.
}

