% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{prob_alive}
\alias{prob_alive}
\title{Survival function for males or females, modelled on patient data until cure_days, then
population data for the remaining time.}
\usage{
prob_alive(time, data, cure_days, boot_coefs, pop_surv_rate, max_age = 100)
}
\arguments{
\item{time}{Time in days.}

\item{data}{Registry dataset of patient cases.}

\item{cure_days}{A number representing time in days corresponding to the cure model assumption for the calculation.}

\item{boot_coefs}{Coefficients for selected bootstrapped dataset.}

\item{pop_surv_rate}{Population survival function in days.}

\item{max_age}{Integer to specify maximum age.}
}
\value{
A survival curve on which to base prevalence predictions.
}
\description{
Survival function for males or females, modelled on patient data until cure_days, then
population data for the remaining time.
}

