% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\docType{data}
\name{registry_data}
\alias{registry_data}
\title{Synthetic patient dataset.}
\format{A data frame with 1908 rows and 8 variables:
\describe{
 \item{ID}{ID, a unique identifier for the patient}
 \item{sex}{sex, "0" for males or "1" for females}
 \item{age}{age, in years}
 \item{DateOfDiag}{date of diagnosis}
 \item{status}{binary variable, 1 if patient is deceased and 0 if alive or censored}
 \item{EventDate}{date of death or censorship}
 \item{stime}{time between date of diagnosis and death or censorship, in days}
 \item{status2}{numeric variable, 1 if deceased and 0 if censored at index date, otherwise NA}
}}
\usage{
registry_data
}
\description{
A dataset synthesised to resemble a disease registry where patient cases are
represented by a unique identifier. Demographic and disease-specific data
required for prevalence estimations are included, such as sex, age, and dates
of diagnosis and death or censorship. Deceased patients have \code{status = 1}, and the
entry in \code{EventDate} is the date of their death. Patients censored at some point
have \code{status = 0}, and the entry in \code{EventDate} is the date of their
censorship. The counted prevalence calculation requires a specified index date after
which all patient cases who are not deceased are censored, in this case 31st August 2013.
The corresponding recoded indicator variable is \code{status2}.
}
\keyword{datasets}

