% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{registry_survival_bootstrapped}
\alias{registry_survival_bootstrapped}
\title{Calculate bootstrap survival coefficients as a way to estimate uncertainty
of the survival model.}
\usage{
registry_survival_bootstrapped(form, data, N_boot = 1000, n_cores = 1)
}
\arguments{
\item{form}{Formula to define survival model by survival time, indicator and covariates.}

\item{data}{Registry dataset of patient cases.}

\item{N_boot}{Integer to specify number of replicates.}
}
\value{
A matrix of coefficients from a Weibull survival model fitted to each bootrapped sample.
}
\description{
Calculate bootstrap survival coefficients as a way to estimate uncertainty
of the survival model.
}
\examples{
form = Surv(time, status) ~ age + sex
data = registry_data
registry_survival_bootstrapped(form, data)
}

