% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{sim_check}
\alias{sim_check}
\title{Estimate consistency of incidence data with a homogeneous Poisson process.}
\usage{
sim_check(data, N_sim = 1e+05)
}
\arguments{
\item{data}{A vector of absolute incidence for each complete year of the registry.}

\item{N_sim}{Number of times to run the simulation.}
}
\value{
Vector of p values for over- and under-dispersion based on the position of the observed sequence variance in the distribution.
}
\description{
This function compares the actual variance of the yearly incidence rates with rates simulated from a Poisson process with overall rate equal to the overall mean rate.
}
\examples{
sim_check(raw_incidence)
}

