% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{smoothed_incidence}
\alias{smoothed_incidence}
\title{Estimate smoothed incidence functions and inspect deviations in the registry data.}
\usage{
smoothed_incidence(data, registry_years, registry_start_year, registry_end_year,
  N = 1000, df = 6)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of the registry data to be used.}

\item{N}{Number of draws from a homogeneous Poisson process.}

\item{df}{Degrees of freedom for the smoothening function.}
}
\value{
Plots of the smoothed incidence function and corresponding deviations.
}
\description{
The first plot shows the cumulative number of diagnoses in the registry by day in black.
This is compared to the red line, which indicates what cumulative diagnosis would look
like if the rate of diagnosis was constant. The green line is a smooth fitted to the
actual cumulative diagnosis data. The second plot shows the deviation of the raw data from
the fitted smooth by day.
}
\details{
The third plot shows the incidence rate per year of the registry, plotted in red. Mean
incidence rate is shown as a dashed black line, with the 95\% confidence interval shown with
dashed blue lines. A smooth fitted to the incidence data is shown in green. In the final
plot, the red line is the deviation of cumulative diagnoses from the fitted smooth (the same
as the black line in plot 2). In order to ascertain if the deviation is within reasonable
bounds, simulation is used. The grey lines represent N draws from a uniform distribution
between 0 and the last day of diagnosis, where N is the number of incident cases, plotted
as deviations from the smooth of cumulative diagnoses. The blue lines are 95\% confidence
intervals drawn from the simulated data.
}
\examples{
smoothed_incidence(load_data(registry_data), registry_years = registry_years,
         registry_start_year = registry_start_year, registry_end_year = registry_end_year)
}

