% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.cincidence}
\alias{summary.cincidence}
\title{Generate a summary of the cumulative_incidence object.}
\usage{
\method{summary}{cincidence}(object, ...)
}
\arguments{
\item{object}{A \code{cumulative_incidence} object.}
}
\description{
Generate a summary of the cumulative_incidence object.
}

