% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.incidence}
\alias{summary.incidence}
\title{Generate a summary of the incidence object.}
\usage{
\method{summary}{incidence}(object, ...)
}
\arguments{
\item{object}{...}
}
\value{
...
}
\description{
Generate a summary of the incidence object.
}
\examples{
...
}

