% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.prevalence}
\alias{summary.prevalence}
\title{Generate a summary of the prevalence object.}
\usage{
\method{summary}{prevalence}(object, ...)
}
\arguments{
\item{object}{A \code{prevalence} object.}
}
\description{
Generate a summary of the prevalence object.
}

