% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.prev.R
\name{summary.survfit.prev}
\alias{summary.survfit.prev}
\title{Summarises survival information at pre-specified years of interest on a
\code{survfit.prev} object.}
\usage{
\method{summary}{survfit.prev}(object, years = c(1, 3, 5), ...)
}
\arguments{
\item{object}{A \code{survfit.prev} object.}

\item{years}{A vector of years at which to estimate survival probability from the bootstrapped
survival curves.}
}
\value{
None, instead displays the survival probabilities to screen as a side-effect.
}
\description{
Survival probability is estimated as the mean of the bootstrapped survival curves at a specific
timepoint, with the 2.5% and 97.5% quantiles providing 95% confidence intervals. Survival probability
can only be estimated at timepoints less than the maximum survival time in the original fitting of
the \code{prevalence} object.
}
\examples{
data(prevsim)

prev_obj <- prevalence(Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate) + event(eventdate),
                       data=prevsim, num_years_to_estimate = c(5, 10), population_size=1e6,
                       start = "2005-09-01",
                       num_reg_years = 8, cure = 5)

survobj <- survfit(prev_obj, newdata=list(age=65, sex=0))

summary(survobj)

summary(survobj, years=c(1, 3, 5, 7))

}

