% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_model}
\alias{survival_model}
\title{Survival function for males or females, modelled on patient data until cure_time, then
population data for the remaining time.}
\usage{
survival_model(time, age, sex, cure_time, boot, daily_survival)
}
\arguments{
\item{time}{Time in days.}

\item{age}{Age of the patient in years.}

\item{sex}{Sex of the patient (0 for males, 1 for females).}

\item{cure_time}{A number representing time in days corresponding to the cure model assumption for the calculation.}

\item{boot}{Selected bootstrapped dataset.}

\item{daily_survival}{Population survival function in days.}
}
\value{
A survival curve on which to base prevalence predictions.
}
\description{
Survival function for males or females, modelled on patient data until cure_time, then
population data for the remaining time.
}

