% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{survival_modelling_diagnostics_sim}
\alias{survival_modelling_diagnostics_sim}
\title{Inspect consistency of survival data between years of the registry and with a Cox Proportional Hazards model.}
\usage{
survival_modelling_diagnostics_sim(form, data, ages, start = NULL,
  num_reg_years = NULL)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{ages}{A vector of ages at which to break the dataset for Kaplan Meier plotting.}

\item{start}{Date from which incident cases are included.}

\item{num_reg_years}{Integer representing the number of complete years of the registry for which incidence is to be calculated.}
}
\value{
A sequence of plots indicated the consistency of survival data between years of the registry and with a Cox Proportional Hazards model.
}
\description{
The first plot is of the Kaplan-Meier survival curve on total cases in the registry. The second plot is the
Kaplan-Meier survival curve for each age group, as delineated by the user using the "ages" argument. The third
plot is of the residuals between the raw data and the fitted Cox Proportional Hazards model. If the model is
a good representation of the data the line should be horizontal. The last plot is of Kaplain Meier survival
curves fitted to cases subdivided by year of diagnosis within the registry (black lines), compared to total
cases shown in blue.
}
\examples{
survival_modelling_diagnostics(Surv(time, status) ~ age(age) + entry(entrydate), registry_data,
ages = c(55, 65, 75, 85, 100), start = "2004-09-01", num_reg_years = 9)
}

