% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{test_homogeneity}
\alias{test_homogeneity}
\title{Summarise disease incidence.}
\usage{
test_homogeneity(entry, year_start = "01-01", truncate_start = FALSE,
  truncate_end = FALSE, population_size = NULL, df = 4,
  proportion = 1e+05, level = 0.95, precision = 2)
}
\arguments{
\item{entry}{Vector of diagnosis dates for each patient in the registry in
the format YYYY-MM-DD.}

\item{year_start}{Date which to use to delimit years in the format MM-DD.
See details for how this is used.}

\item{truncate_start}{See details.}

\item{truncate_end}{See details.}

\item{population_size}{The population of the area covered by the registry. If not provided
then only absolute incidence can be calculated.}

\item{df}{The desired degrees of freedom of the smooth.}

\item{proportion}{The denominator of the incidence rate.}

\item{level}{The desired confidence interval width.}

\item{precision}{The number of decimal places required.}
}
\value{
An S3 object of class \code{incidence} with the following attributes:
  \item{yearly_incidence}{Vector of absolute incidence values for each included
  year of the registry}
  \item{ordered_diagnoses}{Vector of times (days) between diagnosis date and
  the earliest date of inclusion in the registry, ordered shortest to
  longest.}
  \item{smooth}{Smooth fitted to the cumulative incidence data.}
  \item{index_dates}{Dates delimiting the years in which incidence is
  calculated.}
  \item{mean}{List containing absolute yearly incidence as well as relative
  rates.}
  \item{pvals}{p-values resulting to a test of over and under dispersion on the
  incidence data respectively. Used to test the suitability of the homogeneous
  Poission process assumption.}
  \item{dof}{Degrees of freedom of the smooth.}
}
\description{
Calculates incidence by year of the registry data, along with mean incidence
with confidence intervals. A smoothed cumulative incidence function is fit to
the data for inspecting deviations in the registry data from a homogeneous
Poisson process.
}
\details{
TODO Add details about truncation and yearly endpoints
}
\examples{
data(prevsim)

\dontrun{
test_homogeneity(prevsim$entrydate)
}

}
