% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{test_poisson_fit}
\alias{test_poisson_fit}
\title{Estimates consistency of incidence data with an homogeneous Poisson process.}
\usage{
test_poisson_fit(data, N_sim = 1e+05)
}
\arguments{
\item{data}{Vector of absolute incidence values for each included year of the registry
as generated by \code{incidence}.}

\item{N_sim}{Number of simulations to perform.}
}
\value{
Vector of p-values for over- and under-dispersion based on the position of the
observed sequence variance in the distribution.
}
\description{
This function compares the actual variance of the yearly incidence rates with rates
simulated from a Poisson process with overall rate equal to the overall mean rate.
}
\examples{
data(prevsim)

inc <- incidence(prevsim$entrydate)

test_poisson_fit(inc)
}

