% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{test_prevalence_fit}
\alias{test_prevalence_fit}
\title{Chi squared test between prevalence prediction and observed values in the registry.}
\usage{
test_prevalence_fit(object)
}
\arguments{
\item{object}{A \code{prevalence} object.}
}
\value{
P-value from a chi-squared test of difference between prevalence prediction and counted prevalence at the index date.
}
\description{
Chi squared test between prevalence prediction and observed values in the registry.
}
\examples{
data(prevsim)

obj <- prevalence(Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate) + event(eventdate),
                  data=prevsim, num_years_to_estimate = c(5, 10), population_size=1e6,
                  start = "2005-09-01",
                  num_reg_years = 8, cure = 5)

test_prevalence_fit(obj)
}

