% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{data}
\name{UKmortality}
\alias{UKmortality}
\title{General population survival data.}
\format{A data frame with 6666 rows and 4 columns:
\describe{
 \item{calendar_year}{(double) year}
 \item{sex}{(integer) sex encoded as 0 for males and 1 for females}
 \item{age}{(double) age in years}
 \item{rate}{(double) death rate}
}}
\usage{
UKmortality
}
\description{
A dataset containing basic mortality rates stratified by sex and age (yearly from 0 to 100) from
the UK population, obtained from:
\url{http://www.ons.gov.uk/ons/taxonomy/index.html?nscl=Life+Tables#tab-data-tables}.
Adapted from public sector information licensed under the Open Government Licence v3.0.
Data were relabelled according to the mean year of the three-year birth window.
}
\keyword{datasets}

