% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{counted_prevalence}
\alias{counted_prevalence}
\title{Count prevalence from registry data.}
\usage{
counted_prevalence(data, registry_years, registry_start_year, registry_end_year)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal  defining the last year of the registry data to be used.}
}
\value{
A count of prevalence at the index date subdivided by year of diagnosis and inclusion in the registry.
}
\description{
Count prevalence from registry data.
}
\examples{
counted_prevalence(load_data(registry_data), registry_years, registry_start_year, registry_end_year)
}

