% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{counted_prevalence}
\alias{counted_prevalence}
\title{Count prevalence from registry data.}
\usage{
counted_prevalence(entry, eventdate, status, start = NULL,
  num_reg_years = NULL, indexdate = NULL)
}
\arguments{
\item{entry}{Vector of diagnosis dates for each patient in the registry in the format YYYY-MM-DD.}

\item{eventdate}{Vector of dates corresponding to the indicator variable in the format YYYY-MM-DD.}

\item{status}{Vector of binary values indicating if an event has occurred for each patient in the registry. \code{entry}, \code{eventdate}, and \code{status} must all have the same length.}

\item{start}{Date from which incident cases are included, defaults to the earliest date in \code{entry}.}

\item{num_reg_years}{Integer representing the number of complete years of the registry for which incidence is to be calculated, defaults to the number of complete years in the supplied data.}

\item{indexdate}{Index date at which to estimate prevalence, defaults to the date \code{num_reg_years} after \code{start}.}
}
\value{
A vector of length \code{num_reg_years}, representing the corresponding prevalence values.
}
\description{
Count prevalence from registry data.
}
\examples{
data(prevsim)

counted_prevalence(prevsim$entrydate,
                   prevsim$eventdate,
                   prevsim$status)

counted_prevalence(prevsim$entrydate,
                   prevsim$eventdate,
                   prevsim$status,
                   start="2004-01-30", num_reg_years=8,
                   indexdate = "2013-01-30")
}

