% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvcure.R
\name{flexsurvcure_population}
\alias{flexsurvcure_population}
\title{Builds a cure model with an associated population mortality table}
\usage{
flexsurvcure_population(formula, data, dist = "weibull", pop_data = NULL,
  link = "logistic", mixture = T, ...)
}
\arguments{
\item{formula}{A formula expression in conventional R linear modeling
syntax. The response must be a survival object as returned by the
\code{\link{Surv}} function, and any covariates are given on the right-hand
side.  For example,

\code{Surv(time, dead) ~ age + sex}

\code{Surv} objects of \code{type="right"},\code{"counting"},
\code{"interval1"} or \code{"interval2"} are supported, corresponding to
right-censored, left-truncated or interval-censored observations.

If there are no covariates, specify \code{1} on the right hand side, for
example \code{Surv(time, dead) ~ 1}.

By default, covariates are placed on the ``theta'' parameter of the
distribution, representing the cure fraction, through a linear
model with the selected link function.

Covariates can be placed on parameters of the base distribution by using the
name of the parameter as a ``function'' in the formula.  For example, in a
Weibull model, the following expresses the scale parameter in terms of age
and a treatment variable \code{treat}, and the shape parameter in terms of
sex and treatment.

\code{Surv(time, dead) ~ age + treat + shape(sex) + shape(treat)}

However, if the names of the ancillary parameters clash with any real
functions that might be used in formulae (such as \code{I()}, or
\code{factor()}), then those functions will not work in the formula.  A
safer way to model covariates on ancillary parameters is through the
\code{anc} argument to \code{\link{flexsurvreg}}.

\code{\link{survreg}} users should also note that the function
\code{strata()} is ignored, so that any covariates surrounded by
\code{strata()} are applied to the location parameter.}

\item{data}{A data frame in which to find variables supplied in
\code{formula}.  If not given, the variables should be in the working
environment.}

\item{dist}{A string representing one of the built-in distributions of flexsurv.
\code{Surv(time, dead) ~ age + treat, anc = list(shape = ~ sex + treat)}}

\item{pop_data}{A data frame comprising population survival as a daily rate for 36525 days (100 years).
It can also be stratified by other variables that are found in the survival \code{formula} for this model,
such as sex.}

\item{link}{A string representing the link function to use for estimation of the
cure fraction.  Defaults to logistic.}

\item{mixture}{optional TRUE/FALSE to specify whether a mixture model should be fitted.  Defaults to TRUE.}

\item{...}{other arguments to be passed to \code{\link{flexsurvreg}}.}
}
\description{
Builds a cure model with an associated population mortality table
}
