% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{functional_form_age}
\alias{functional_form_age}
\title{Inspect functional form of age.}
\usage{
functional_form_age(form, data, df = 4)
}
\arguments{
\item{form}{...}

\item{data}{A registry dataset of patient cases.}

\item{df}{Degrees of freedom for the smooth.}
}
\value{
Plots of the functional form of age.
}
\description{
Inspect functional form of age.
}
\examples{
functional_form_age(registry_data_r)
}

