% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{functional_form_age}
\alias{functional_form_age}
\title{Inspect functional form of age.}
\usage{
functional_form_age(data, registry_years, registry_start_year,
  registry_end_year, df = 4)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of the registry data to be used.}

\item{df}{Degrees of freedom for the smooth.}
}
\value{
Plots of the functional form of age.
}
\description{
Inspect functional form of age.
}
\examples{
functional_form_age(load_data(registry_data), registry_years, registry_start_year = registry_start_year, registry_end_year = registry_end_year)
}

