% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\title{Calculate absolute incidence from registry data.}
\usage{
incidence(entry, start = NULL, num_reg_years = NULL)
}
\arguments{
\item{entry}{(character): Vector of diagnosis dates for each patient in the registry in the format YYYY-MM-DD.}

\item{start}{(character): Date from which incident cases are included in the format YYYY-MM-DD, defaults to the earliest entry date.}

\item{num_reg_years}{(integer): The number of years of the registry for which incidence is to be calculated, defaults to using all available complete years.}
}
\value{
Vector of length num_reg_years of doubles, representing the number of absolute incidence values for each included year of the registry.
}
\description{
Calculate absolute incidence from registry data.
}
\examples{
incidence(prevsim$entrydate, start="2004-01-01", 8)
}

