% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{incidence_age_distribution}
\alias{incidence_age_distribution}
\title{Plot the age distribution of incident cases in the registry data.}
\usage{
incidence_age_distribution(agedata, df = 10)
}
\arguments{
\item{agedata}{Vector of ages at diagnosis for each patient in the registry.}

\item{df}{Degrees of freedom for the smooth.}
}
\value{
Plot of the raw data and smoothed age distribution function.
}
\description{
Plot the age distribution of incident cases in the registry data.
}
\examples{
incidence_age_distribution(prevsim_r$age)
}

