% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_plots.R
\name{inspect_incidence}
\alias{inspect_incidence}
\title{Inspect incidence rate.}
\usage{
inspect_incidence(object, level = 0.95)
}
\arguments{
\item{object}{A \code{cincidence} object.}

\item{level}{Double representing the desired confidence interval width.}
}
\value{
Plot of incidence rate, confidence interval and smoothed incidence function.
inspect_incidence(c_inc)
}
\description{
This function generates a plot from the cumulative incidence object. The incidence rate per
year of the registry is shown in red. Mean incidence rate is shown as a dashed black line,
with the 95\% confidence interval shown with dashed blue lines. The smooth fitted to the
cumulative incidence data is shown in green.
}

