% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{n_year_estimates}
\alias{n_year_estimates}
\title{Calculate predicted prevalence for a given number of years.}
\usage{
n_year_estimates(object, num_years_to_estimate, population_size, level = 0.95,
  precision = 2)
}
\arguments{
\item{object}{Object returned from prevalence\(\).}

\item{num_years_to_estimate}{Integer representing number of years prevalence is to be calculated for.}

\item{population_size}{Integer corresponding to the size of the population.}

\item{level}{Double representing the desired confidence interval width.}

\item{precision}{Integer representing the number of decimal places required.}
}
\value{
An estimate of prevalence, in total and /100,000 with confidence intervals.
}
\description{
Calculate predicted prevalence for a given number of years.
}
\examples{
n_year_estimates(prevalence_total, num_years_to_estimate = 3, population_size = 1700000)
}

