% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Predict prevalence for a given number of years.}
\usage{
prevalence(data, registry_years, registry_start_year, registry_end_year,
  N_years, daily_survival_males, daily_survival_females, cure_time,
  N_boot = 1000, Max_Yearly_Incidence = 500)
}
\arguments{
\item{data}{A registry dataset of patient cases generated using load_data().}

\item{registry_years}{A vector of dates delineating years of the registry.}

\item{registry_start_year}{Ordinal defining the first year of the registry data to be used.}

\item{registry_end_year}{Ordinal defining the last year of the registry data to be used.}

\item{N_years}{Number of years to model.}

\item{daily_survival_males}{Population survival function (Males).}

\item{daily_survival_females}{Population survival function (Females).}

\item{cure_time}{Cure model assumption for the calculation.}

\item{N_boot}{Number of bootstrapped calculations to perform.}

\item{Max_Yearly_Incidence}{A number larger than the expected yearly incidence
to allow for variation in incidence between years.}

\item{age_division}{Can be "None", "Adult" or "Paediatric".}

\item{age_cut}{A number, in years, to specify the cut-off between adult and
paediatric cases.}

\item{sex}{Can be "Both", "Males" or "Females".}
}
\value{
Predicted number of prevalent cases for each year, the age distribution of the
       prevalent cases, an indication of sampling variation and raw incidence data to
       allow for crosschecking with previous calculations.
}
\description{
Predict prevalence for a given number of years.
}
\examples{
by_year_total <- prevalence(load_data(registry_data), registry_years, registry_start_year,
                           registry_end_year, daily_survival_males = daily_survival_males,
                           daily_survival_females = daily_survival_females, cure_time = cure*365,
                           N_years = 10)
}

