% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Estimates point prevalence at a specific index date using a combination of available registry
data, and Monte Carlo simulated incident data.}
\usage{
prevalence(form, data, num_years_to_estimate, population_size, start = NULL,
  num_reg_years = NULL, cure = 10, N_boot = 1000,
  max_yearly_incidence = 500, level = 0.95, precision = 2,
  proportion = 1e+05, population_data = NULL, n_cores = 1)
}
\arguments{
\item{form}{Formula where the LHS is represented by a standard \code{Surv} object, and the RHS has three special function arguments:
\code{age}, the column where age is located;
\code{sex}, the column where sex is located;
\code{entry}, the column where dates of entry to the registry are located; and
\code{event}, the column where event dates are located.

This formula is used in the following way:

\code{Surv(time, status) ~ age(age_column_name) + sex(sex_column_name) + entry(entry_column_name) + event(event_column_name)}

Using the supplied \code{prevsim} dataset, it is therefore called with:

\code{Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate) + event(eventdate)}}

\item{data}{A data frame with the corresponding column names provided in \code{form}.}

\item{num_years_to_estimate}{Vector of integers representing number of years of data to consider when
estimating point prevalence. If this value is greater than \code{num_registry_years},
incident cases for the difference will be simulated.}

\item{population_size}{Integer corresponding to the size of the population.}

\item{start}{Date from which incident cases in the registry data are to be included in the prevalence estimation, in
the format YYYY-MM-DD. Defaults to the earliest incident case in the supplied registry.}

\item{num_reg_years}{Integer representing the number of complete years of the registry for which incidence is to be calculated.
Defaults to the number of complete years of registry data. Note that if more registry years are supplied than the number of years
to estimate prevalence for, the survival data from the surplus registry years are still involved in the survival model fitting.}

\item{cure}{Integer defining cure model assumption for the calculation (in years). A patient who has survived beyond the cure time
has their survival probability incorporating mortality rates of the underlying population.}

\item{N_boot}{Integer representing number of bootstrapped calculations to perform.}

\item{max_yearly_incidence}{Integer larger than the expected yearly incidence
to allow for variation in incidence between years.}

\item{level}{Double representing the desired confidence interval width.}

\item{precision}{Integer representing the number of decimal places required.}

\item{proportion}{The population ratio to estimate prevalence for.}

\item{population_data}{A dataframe which must contain the columns \code{age}, \code{rate}, and \code{sex},
where each row is the mortality rate for a person of that age and gender. Ideally, age ranges from [0, 100].
Defaults to the supplied data \code{UKmortality}; see the documentation for this dataframe for help in
providing custom ones.}

\item{num_cores}{The number of CPU cores to run the fitting of the bootstrapped survival models across.
Defaults to 1; multi-core functionality is provided by the \code{doParallel} package.}
}
\value{
An S3 object of class \code{prevalence} with the following attributes:
\item{estimates}{Estimated prevalence at the index date for each of the years in \code{num_years_to_estimate}.}
\item{simulated}{A list containing items related to the simulation of prevalence contributions, see \code{prevalence_simulated}}.
\item{counted}{Contributions to prevalence from each of the supplied registry years, see \code{prevalence_counted}.}
\item{start_date}{The starting date of the registry data included in the estimation.}
\item{index_date}{The index date at which the point prevalence was calculated for.}
\item{known_inc_rate}{The known incidence rate for years included in the registry.}
\item{nregyears}{Number of years of registry data that were used.}
\item{nbootstraps}{The number of bootstrap survival models fitted during the calculation.}
\item{pval}{The p-value resulting from the chi-square test between the simulated and counted prevalent cases for the years of
registry data available.}
\item{y}{The Surv object used as the response in the survival modeling.}
\item{means}{The co-variate means from the data.}
}
\description{
The most important parameter is \code{num_years_to_estimate}, which governs the number of
previous years of data to use when estimating the prevalence at the index date. If this
parameter is greater than the number of years of known incident cases available in the
supplied registry data (specified with argument \code{num_registry_years}), then the
remaining \code{num_years_to_estimate - num_registry_years} years of incident data
will be simulated using Monte Carlo simulation.
}
\details{
The larger \code{num_years_to_estimate}, the more accurate the prevalence estimate will be,
provided an adequate survival model can be fitted to the registry data. It is therefore
important to provide as much clean registry data as possible.

Simulated cases are marked with age and sex to enable agreement with population survival
data where a cure model is used, and calculation of the posterior distributions of each.
}
\examples{
data(prevsim)

prevalence(Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate) + event(eventdate),
           data=prevsim, num_years_to_estimate = c(5, 10), population_size=1e6,
           start = "2005-09-01",
           num_reg_years = 8, cure = 5)

prevalence(Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate) + event(eventdate),
           data=prevsim, num_years_to_estimate = 5, population_size=1e6)

prevalence(Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate) + event(eventdate),
           data=prevsim, num_years_to_estimate = c(3,5,7), population_size=1e6, n_cores=4)

}

