% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Predict prevalence for a given number of years.}
\usage{
prevalence(form, data, N_years, cure_time = NULL, start = NULL,
  num_reg_years = NULL, N_boot = 1000, max_yearly_incidence = 500,
  population_data = NULL, n_cores = 1)
}
\arguments{
\item{form}{Formula where the LHS is represented by a standard \code{Surv} object, and the RHS has three special function arguments:
\code{age}, the column where age is located;
\code{sex}, the column where sex is located; and
\code{entry}, the column where dates of entry to the registry are located.

This formula is used in the following way \code{Surv(time, status) ~ age(age_column_name) + sex(sex_column_name) + entry(entry_column_name)}

Using the supplied \code{prevsim} dataset, it is therefore called with:

\code{Surv(time, status) ~ age(age) + sex(sex) + entry(entrydate)}}

\item{data}{Dataset of patient cases generated by load_data().}

\item{N_years}{Integer representing number of years to model.}

\item{cure_time}{Integer defining cure model assumption for the calculation.}

\item{start}{Date from which incident cases are included.}

\item{num_reg_years}{Integer representing the number of complete years of the registry for which incidence is to be calculated.}

\item{N_boot}{Integer for number of bootstrapped calculations to perform.}

\item{max_yearly_incidence}{Integer larger than the expected yearly incidence
to allow for variation in incidence between years.}

\item{pop_vers}{...}

\item{colnames}{Vector of column names indicating required variables: age, sex, entry_date, status and time.}
}
\value{
Predicted number of prevalent cases for each year, the age distribution of the
       prevalent cases, an indication of sampling variation and raw incidence data to
       allow for crosschecking with previous calculations.
}
\description{
Predict prevalence for a given number of years.
}
\examples{
names = list(age = 'age', sex = 'sex', entry_date = 'DateOfDiag', status = 'status', time ='stime')
prevalence_total <- prevalence(registry_data,
                               N_years = 10,
                               cure_time = 10*365,
                               start = "2005-09-01",
                               num_reg_years = 8,
                               colnames = names)

}

