% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence_by_age}
\alias{prevalence_by_age}
\title{Output extrapolated number of prevalent cases for specified age bands.}
\usage{
prevalence_by_age(object, age_intervals = seq(10, 80, by = 10))
}
\arguments{
\item{object}{Object returned by prevalence\(\) from which to extrapolate.}

\item{age_intervals}{Vector of integers to delineate age bands.}
}
\value{
Vector of predicted number of prevalent cases for specified age bands.
}
\description{
Output extrapolated number of prevalent cases for specified age bands.
}
\examples{
prevalence_by_age(dist = post_age_dist_male, registry_end_year = 4,
                 num_years_to_estimate = num_years_to_estimate)
}

