% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{raw_incidence}
\alias{raw_incidence}
\title{Disease incidence.}
\usage{
raw_incidence(entry, start = NULL, num_reg_years = NULL)
}
\arguments{
\item{entry}{Vector of diagnosis dates for each patient in the registry in
the format YYYY-MM-DD.}

\item{start}{Date from which incident cases are included in the format
YYYY-MM-DD. Defaults to the earliest entry date.}

\item{num_reg_years}{The number of years of the registry for which incidence
is to be calculated. Defaults to using all available complete years.}
}
\value{
Vector of length num_reg_years of integers, representing the number
  of absolute incidence values for each included year of the registry.
}
\description{
Calculates yearly incidence for the available registry data.
}
\examples{
data(prevsim)

raw_incidence(prevsim$entrydate, start="2004-01-01", num_reg_years=8)
raw_incidence(prevsim$entrydate)
raw_incidence(prevsim$entrydate, start="2005-05-01", num_reg_years=5)
raw_incidence(prevsim$entrydate, start="2005-05-01")

}
\seealso{
Other incidence functions: \code{\link{incidence_age_distribution}},
  \code{\link{incidence}},
  \code{\link{mean_incidence_rate}},
  \code{\link{plot.incidence}},
  \code{\link{test_incidence_fit}}
}

