% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{registry_survival_bootstrapped}
\alias{registry_survival_bootstrapped}
\title{Calculate bootstrap survival coefficients as a way to estimate uncertainty
of the survival model.}
\usage{
registry_survival_bootstrapped(data, N_boot = 1000)
}
\arguments{
\item{data}{Dataset of patient cases generated by load_data(), restricted to the first complete year
of the registry.}

\item{N_boot}{Number of replicates.}
}
\value{
The coefficients from a Weibull survival model fitted to each bootrapped sample.
}
\description{
Calculate bootstrap survival coefficients as a way to estimate uncertainty
of the survival model.
}
\examples{
registry_survival_bootstrapped(data, N_boot = 1000)
}

