% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_prevalence.R
\name{sim_prevalence}
\alias{sim_prevalence}
\title{Estimate prevalence using Monte Carlo simulation.}
\usage{
sim_prevalence(data, index, starting_date, inc_model, surv_model,
  age_column = "age", N_boot = 1000, n_cores = 1)
}
\arguments{
\item{data}{A data frame with the corresponding column names provided in
\code{form}.}

\item{index}{The date at which to estimate point prevalence as a string in the format
YYYY-MM-DD.}

\item{starting_date}{The initial date to start simulating prevalence from as a \code{Date} object.
Typically the index date - (Nyears * 365.25). Allows for non-whole year prevalence estimations.}

\item{inc_model}{An object that has a \code{draw_incident_population}
method. See the vignette for further guidance.}

\item{surv_model}{An object that has a \code{predict_survival_probability}
method. See the vignette for further guidance.}

\item{age_column}{A string providing the name of the column that holds patient age. If provided
then patients are assumed to die at 100 years.}

\item{N_boot}{Number of bootstrapped calculations to perform.}

\item{n_cores}{Number of CPU cores to run the fitting of the bootstrapped
survival models. Defaults to 1; multi-core functionality is provided by the
\code{doParallel} package.}
}
\value{
A list with the following attributes:
  \item{mean_yearly_contributions}{A vector of length
  \code{num_years_to_estimate}, representing the average number of prevalent
  cases subdivided by year of diagnosis across each bootstrap iteration.}
  \item{posterior_age}{Posterior distributions of age, sampled at every
  bootstrap iteration.} \item{yearly_contributions}{Total simulated prevalent
  cases from every bootstrapped sample.} \item{pop_mortality}{Population
  survival rates in the format of a list, stratified by sex.}
  \item{nbootstraps}{Number of bootstrapped samples used in the prevalence
  estimation.} \item{coefs}{The bootstrapped Weibull coefficients used by the
  survival models.} \item{full_coefs}{The Weibull coefficients from a model
  fitted to the full dataset.}
}
\description{
Estimates prevalent cases at a specific index date by use of Monte Carlo
simulation. Simulated cases are marked with age and sex to enable agreement
with population survival data where a cure model is used, and calculation of
the posterior distributions of each.
}
\examples{
data(prevsim)

\dontrun{
# TODO Generate incidence and survival object
# TODO Do I actually want to keep this and the counted function
# as exports?
simulated_prevalence(prevsim, "2013-01-01",
                     starting_date="2003-01-01",
                     inc_model, surv_model)
}

}
\seealso{
Other prevalence functions: \code{\link{counted_prevalence}},
  \code{\link{prevalence_counted_legacy}},
  \code{\link{prevalence_simulated_legacy}},
  \code{\link{prevalence}},
  \code{\link{test_prevalence_fit_legacy}},
  \code{\link{test_prevalence_fit}}
}
