% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKmortality}
\alias{UKmortality}
\title{General population survival data.}
\format{A data frame with 6666 rows and 4 columns:
\describe{
 \item{calendar_year}{year}
 \item{sex}{boolean variable; 0 for males and 1 for females}
 \item{age}{age in years}
 \item{rate}{death rate}
}}
\usage{
UKmortality
}
\description{
A dataset containing basic mortality rates stratified by sex and age (0 to 100 year) from
the UK population, yearly, obtained from:
\url{http://www.ons.gov.uk/ons/taxonomy/index.html?nscl=Life+Tables#tab-data-tables}.
Adapted from public sector information licensed under the Open Government Licence v3.0.
Data were relabelled according to the mean year of the three-year birth window.
}
\keyword{datasets}

